+++ spec/active_interaction/integration/interface_interaction_spec.rb
@@ -1,0 +1,12 @@
+# coding: utf-8
+
+require 'spec_helper'
+require 'json'
+require 'yaml'
+
+describe 'InterfaceInteraction' do
+  it_behaves_like 'an interaction',
+    :interface,
+    -> { [JSON, YAML].sample },
+    methods: [:dump, :load]
+end